requirements.txt
Flask
gunicorn #for Heroku
python-dotenv
requests